requirements.txt
flask
sqlalchemy
psycopg2-binary
python-dotenv
requests
Flask-SQLAlchemy
pytest